First version of the writing tool, works somewhat though far from finished and not exactly usable. Need better UI elements, more tools in the word processor, there's a few issues which need to be ironed out.
To do:
Full redesign of the project structure & change of the UI for the dashboard, make the word processor work better, more coherent & easy.
Rework of the story-wiki, and perhaps add a history to show past edits as well as a format system for chapters & other elements. To be seen.